reorganize project folder
