+++ js/data/test/getTime.test.js
@@ -1,4 +1,4 @@
-import Timer from '../getTime';
+import Timer from '../../service/getTime';
 import {assert} from 'chai';
 
 suite(`getTime`, () => {
