+++ js/screens/result.js
@@ -3,7 +3,6 @@ import resultTemplate from '../service/resultService';
 import {resetGame} from '../service/gameService';
 
 export default () => {
-
   const screenContainer = getElementFromTemplate(`
   <section class="main main--result">
     <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
