+++ js/service/getTextResult.js
